+++ ChangeLog.md
@@ -1,4 +1,15 @@
 # ChangeLog
+### <u>11/4/25</u>
++ added support for method declaration
++ added support for multiple inheritance
++ added support for abstract with inheritance
++ added override for abstract methods (import automatically detects python version)
++ added Enum integration
++ added choice between merging common properties between parents or having distinct properties for each constructor
++ added better exception handling and debug toggle
++ log toggle generates a log file
+### <u>03/4/25</u>
++ modified camelize() to accomodate for snake_case input
 ### <u>31/3/25</u>
 + added abstract class support
 + added changelog